ui: Move Statements views
Release note: None
