+++ pkg/ui/modules/statements/views/barCharts.tsx
