+++ pkg/ui/modules/statements/views/diagnostics/activateDiagnosticsModal.tsx
