+++ pkg/ui/modules/statements/views/diagnostics/diagnosticStatusBadge.styl
