+++ pkg/ui/modules/statements/views/diagnostics/diagnosticStatusBadge.tsx
