+++ pkg/ui/modules/statements/views/diagnostics/diagnosticsUtils.ts
