+++ pkg/ui/modules/statements/views/diagnostics/diagnosticsView.spec.tsx
