+++ pkg/ui/modules/statements/views/diagnostics/diagnosticsView.tsx
