+++ pkg/ui/modules/statements/views/planView.spec.tsx
@@ -11,7 +11,7 @@
 import { assert } from "chai";
 
 import { cockroach } from "src/js/protos";
-import { FlatPlanNode, FlatPlanNodeAttribute, flattenTree, flattenAttributes } from "src/views/statements/planView";
+import { FlatPlanNode, FlatPlanNodeAttribute, flattenTree, flattenAttributes } from "./planView";
 import IAttr = cockroach.sql.ExplainTreePlanNode.IAttr;
 import IExplainTreePlanNode = cockroach.sql.IExplainTreePlanNode;
