+++ pkg/ui/modules/statements/views/statementDetails.tsx
@@ -37,8 +37,8 @@ import { SortSetting } from "src/views/shared/components/sortabletable";
 import SqlBox from "src/views/shared/components/sql/box";
 import { formatNumberForDisplay } from "src/views/shared/components/summaryBar";
 import { ToolTipWrapper } from "src/views/shared/components/toolTip";
-import { PlanView } from "src/views/statements/planView";
-import { SummaryCard } from "../shared/components/summaryCard";
+import { PlanView } from "./planView";
+import { SummaryCard } from "src/views/shared/components/summaryCard";
 import { approximify, latencyBreakdown, longToInt, rowsBreakdown } from "./barCharts";
 import { AggregateStatistics, makeNodesColumns, StatementsSortedTable } from "./statementsTable";
 import { getMatchParamByName } from "src/util/query";
